src/copilot-chat/src/pyproject.toml
[project]
name = "copilot-chat"
description = "CoPilot."
authors = [
    { name = "Lei Qu", email = "lequ@microsoft.com" }
]
dependencies = [
    "colorama>=0.4.6",
    "faiss-cpu>=1.9.0",
    "transformers>=4.46.1",
    "langchain_community>=0.3.4",
    "matplotlib>=3.9.2",
    "scipy>=1.11.1",
    "scikit-learn>=1.5.2",
    "termcolor>=2.5.0",
    "requests>=2.32.3",
    "eval_type_backport>=0.2.2",
    "requests"
]
# Align Python version to Docker image provided Python version.
# docker run --rm mcr.microsoft.com/cbl-mariner/base/python:3 python3 -V
requires-python = "==3.9.*"
classifiers = ["Private :: Do Not Upload"]
# Generate version dynamically with versioningit.
dynamic = ["version"]


[tool.rye.scripts]
copilot-agent = "python -m copilot_agent"

[build-system]
requires = ["hatchling", "versioningit"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
universal = true
dev-dependencies = [
    "pandas>=2.2.2",
    "azure-kusto-data>=4.5.1",
    "flask",
    "openai",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
source = "versioningit"

[tool.versioningit]
default-version = "0.0.0+unknown"

[tool.versioningit.vcs]
method = "git"
match = ["v*"]

[tool.versioningit.format]
distance = "{next_version}-dev{distance}+{rev}"
dirty = "{version}+dirty"
distance-dirty = "{next_version}-dev{distance}+{rev}.dirty"

[tool.versioningit.template-fields.version-tuple]
pep440 = true
double-quote = false

[tool.versioningit.write]
file = "_version.py"
template = """\
# This file is automatically generated by versioningit.
__version__ = '{version}'
__version_tuple__ = {version_tuple}\
"""

[tool.hatch.build]
# Force hatch to include files even they are excluded in .gitignore file.
artifacts = ["src/copilot_agent/_version.py"]
include = ["data/**", "src/copilot_agent/prompts/**"]

[tool.hatch.build.targets.wheel]
# Necessary when package name is different from project name.
packages = ["copilot_agent"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
extend-select = [
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "FA",   # flake8-future-annotations
    "Q",    # flake8-quotes
    "PGH",  # pygrep-hooks
    "RUF",  # ruff
    "W",    # pycodestyle
    "D",    # pydocstyle
    "UP",   # pyupgrade
    "YTT",  # flake8-2020
    "C90",  # mccabe
]

[tool.ruff.lint.per-file-ignores]
# Ignore all D (docstring) rules in the test folder
"**/{tests,docs,tools}/*" = [
    "D100",  # Missing docstring in public module
    "D101",  # Missing docstring in public class
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
    "D104",  # Missing docstring in public package
    "D105",  # Missing docstring in magic method
    "D106",  # Missing docstring in public nested class
    "D107",  # Missing docstring in __init__
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.lint.isort]
known-first-party = [
]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true
quote-style = "single"
